refactor!: rename to feline-theme
