+++ lua/feline-theme/example/components.lua
