+++ lua/feline-theme/example/theme.lua
