+++ lua/feline-theme/schema/colors.lua
@@ -1,4 +1,4 @@
-local feline = require('compline.schema.feline')
+local feline = require('feline-theme.schema.feline')
 
 local M = {}
